requirements.txt
# RaspberryPi
RPi.GPIO

# Flask
Flask==0.10.1
MarkupSafe==0.18
Werkzeug==0.9.4
itsdangerous==0.23
Jinja2==2.7.1

# API
marshmallow>=0.4.1
Flask-Classy==0.6.8

# Testing
nose==1.3.0
Flask-Testing==0.4